In posterFind(), do not return ERROR if the wanted poster is found in cache, but the cached POSTER_STR points to a closed poster. Instead, just remove the POSTER_STR from allPosters and go on with the normal find procedure.
This fixes the situation where a given poster is created/deleted and
created back, but clients cannot find the new poster because they
already have the old one in cache.

NOTE: we cannot free() the POSTER_STR structure since the client program
may still have a POSTER_ID that points to it.