Reorganize the way taskFromThread() is implemented.
There are strange side effects with the pseudo-varargs games played
in taskSpawn() that arise whith the old version with gcc -O2 and cause
stack corruption and segfaults in tclserv.

At some point all this interface will need to be changed (breaking
both source and binary compatibility) to use <stdarg.h>. But for now
this should be ok.