Implement h2-semaphores with POSIX semaphores
Added an option in autoconf script to select the preferred
implementation (SVr4 or POSIX). Note that the POSIX implementation
modifies the h2Devs and is not binary compatible with older versions.
The public API is not changed.

The major change is the inclusion of an array of MAX_SEM sem_t into the
h2Devs, making them grow from 64 to 320 bytes. The global h2Devs array
is now about 45kB, but this could be reduced by tuning the MAX_SEM
value (currently 20).

Since POSIX semaphores are not mutexes and there is no API to set the
value of posix sems, the h2semGive() function tries to prevent the
semaphore value to grow above 1, but there is still a possible race if
two threads do h2semGive() on the same semaphore at the same time.