Fix h2semTake() when called with a timeout under ART-linux
h2semTake() with a timeout is implemented via a watchdog. This
currently relies on SIGALRM being triggered by sysLib at every tick.

Under ART-linux, sysLib doesn't use any signal, so we have to generate
one in order to interrupt sem_wait() or semop(). We use the watchdog
to send SIGALRM via the pthread_kill() function and then check the
result of sem_wait() or semop() as before. Before that, the target
thread is prepared to receive the signal with sigaction().

Note: Linux implements sem_timedwait() and semtimedop() for that
purpose but they are not standard and not implemented on other
systems.