unblock the ALRM signal explicitely in watchdoc functions.
Previously we restored the saved signal mask using
sigprocmask(SIG_SET,...) but on Darwin this doesn't seem to work
(the signal stays blocked for all threads). So instead now
use SIG_UNBLOCK to unblock it explicitely.

This makes the h2timersem regress test work (and hopefully also modules
that do posterFind() on Darwin.