Add FIO_RESIZE to the posterIoctl() function
As the name suggests, FIO_RESIZE resizes an existing poster. It takes a
size_t * parameter that should point to the new desired size. A new regression
test has been added.

Do not use for hard real-time software, as this implies non deterministic
memory allocation with the current smMemLib() implementation. In all cases,
calling this ioctl implies copying the whole content of the old poster data to
the new (resized) location (this could be improved by optimizing smMemRealloc()
function).

As a side effect, the posterAddr() result cannot be cached between several
posterTake() invocation: the actual addr must be refetched each time. As long
as posters are not resized, any caching that may be done in legacy software
will still work, though.

The remote posterLib() behaves correctly with respect to resizing: a
remote posterTake() will fetch the content of the whole poster and give the
expected result. This may imply resizing the local cache, and thus may fail.

posterRead() and posterWrite() may now return fewer bytes than requested, if
the poster was resized to a smaller buffer (this was already accounted for in
the API).