Implement h2rngRealloc() that resizes an existing ring buffer
A new ring buffer is created and old content is copied in the new buffer.
The old id becomes invalid after the resize. Resizing to a size smaller than
current buffer content is not allowed.

Note: for now, shrinking is explicitly disallowed (even if the current content
would fit), to avoid breaking existing code that may not be prepared to deal
with this. The code allowing a shrink is simply commented out, so that it is
clear that this is forbidden purposedly.

Added a simple test for h2rngRealloc() in regress.