Implement csMboxUpdate() and gcomUpdate()
These functions resize existing mailboxes for a task (requests or reply
mailboxes, or both). This is transparent for the clients and the contents of
the mailboxes is preserved. This may fail if there is not enough shared memory
for the operation (in case of a failure, these operations are a noop).

Shrinking a mailbox is discouraged, and currently explicitly disallowed, since
this may trigger unexpected failures for clients willing to send a message
that does not fit anymore in the new mailbox (that's of course expected, but
legacy code may not be prepared to deal with this).

Expanding a mailbox should not raise any issue, as long as no client does
horrid stuff like caching internal mbox pointers or the like.