Implement mutex semaphores with a ticket lock.
This provides FIFO semantics for locked threads and achieves better
equity among threads. This is noteworthy in the case of the commonStruct
used to store the Genom IDS. If an execution task is looping, the
control task will never manage to lock the IDS to run.
With the ticket lock, locked tasks will be unlocked in a fifo order.

XXX to match the VxWorks SEM_Q_FIFO behaviour , ther other types of
semaphores should also implement this.