Remember base address of smMemLib separately from free blocklist addr
The list of free blocks usually starts at the base address of the mapped
shared memory segement, but this is not guaranteed. For instance, if all
memory is allocated, the free block list will become NULL. If a block is
later freed, the free block list address will be in the middle of the
shared memory segment.

Since the base address returned by smMemBase() is used in
smObjLocaltoGobal() and smObjGlobalToLocal() to store the global
addresses of h2devs, it must never change during the lifetime of a
process.

This commit introduces a new smMemBaseAddr global variable that
remembers the shmat() address and never changes between smMemInit() and
smMemEnd().