Don't use sigprocmask(2) in wdLib
sigprocmask(2) is unspecified in a multi-threaded
environment. pthread_sigmask(3) must be used instead.

This is still not sufficient though, because SIGALRM should be blocked for all
threads in order to actually perform mutual exclusion with the sysClkInt()
SIGALRM handler. Added an XXX note about this in the code.

This happens to fix #117, though, because it appears that the sigprocmask()
call in wdStart() is making MacOS X sigwait(3) return without error but with
errno set to EINTR, while pthread_sigmask(3) does not have this effect.