Set errno and print error if a RPC call fails in remotePosterTake/Give()
The errno is set to S_remotePosterLib_BAD_RPC, as in other functions, if the
RPC call to poster_read_1() or poster_write_1() fails. This avoids upper layers
to print "unknown error".

While here, also print to stderr the actual RPC error in remotePosterTake(), as
what is done in remotePosterGive().