Don't do a remote posterRead() in remotePosterTake(POSTER_WRITE)
The remote posterRead() is mandatory for remotePosterTake(POSTER_READ).
However, for a POSTER_WRITE operation, there is no need to update it,
because POSTER_WRITE (and posterResize() FWIW) can only be called by
one process, that is necessarily on the "remote side" of the network
since we are calling remotePosterTake(). Since no remote component can
update the poster, the cache is up-to-date by construction.