Fix a race when allocating h2sem
When a new page of semaphores is needed, we used to release the global
h2dev semaphore to call h2devAloc() and re-lock it once the new device
was allocated. This left a window for other tasks wanting to allocate
a semaphore to step in and stumble on an error because the semaphore
page associated with the new device hasn't be initialized yet.

Instead, introduce an h2devAllocUnlocked() function (that is supposed
to already hold the global h2devs lock) and use it to close the race.

This was uncovered by the csLibMax regress test.

XXX while there, accidentally clean-up white space.