h2devEnd: delay the destruction of the 1st page of semaphores.
This insures that the global h2dev semaphore (sem 0 in the 1st page)
will stay around until all its users are done.  When semaphores are
destroyed, the 1st page is skipped.  It is destroyed explicitely with
a new h2semDelete0() function later.

Avoids random errors during h2devEnd() that were not catched but could
lead to inconsistent states and undefined behaviour.