Fix posterServ segfault when reading an empty poster (or any error)
Using the unsigned res->data_len to store the result of localPosterRead() is
wrong when localPosterRead() returns a negative ERROR: this makes the length
equal to INT_MAX and xdr_array routine scans unallocated memory.

Use a local int to check the error and update res->data_len with len only in
case of success. Set it to 0 in case of error to prevent xdr_array from reading
uninitialized memory.