If smMemAttach() fails in shmat(2), don't destroy the memory segment
shmat(2) may fail to map the shared memory segment, but this does not
mean that the memory segment must be destroyed. Indeed, this will break
any other running pocolibs program that was successful in doing the mapping.

For the same reason, the SM_MEM h2Dev must not be freed.

This error handling was probably pasted from smMemInit(), which creates the
shared memory segment and must thus clean everything in case of error.

ok matthieu@