Add a generation number to the h2dev number
An h2dev numer now contains additional bits (currently 12) in the MSB, so
that mboxes, posters or tasks that would be deleted and recreated with the
same type/uid cannot be accessed by an old (cached) h2dev number.

Without this, an old h2dev number would still allow access a newly recreated
device, most probably with an undefined behaviour if the new device is
unrelated to the previous old one.

The API is unchanged but the H2DEV_STR is modified, so all dependent software
must be recompiled to be able to access the new structure.

Bump comLib major to 12.
OK matthieu@