Update compile-time test for working posix timers
There is an issue spotted on ubuntu-22.04 and 24.04, as well as other
distributions, that affects posix timer signal delivrery. It is easily
triggered by spawning and terminating threads quickly while the timer is
running. At some point, the timer signal will stop being delivered.
Inspecting the timer with timer_gettime(2) shows it still counting,
reconfiguring the timer with timer_settime(2) makes it work again and
manually sending the timer signal also always work.

This seems to affect kernels from 6.4 to 6.11 inclusive, i.e. only 6.3 and
earlier are immune to the issue at this time.

In practice, this issue is triggering in existing sofware doing some
aio(7) work and thus involving thread creation/destruction. So a new test is
added in configure script to (hopefully) detect the issue at compile time and
disable the use of posix timers in favor of setitimer(2).