Fix potential truncation of default task name in taskLibInit()
pid_t is nowadays 32 bits long on Linux, thus can be 10 digits long. With the
'tUnix" prefix, the default task name in taskLibInit() may thus use up to 16
chars.

Using a buffer of length 12 may truncate the name and later lead to a mix
between unrelated tasks (associated to pid differing only in the last digits)
in mboxInit().